+++ src/Vector/DBC/tests/unittests/test_Regression.cpp
@@ -39,22 +39,35 @@ BOOST_AUTO_TEST_CASE(Regression)
     std::cout << "Input file: " << infilename << std::endl;
     BOOST_REQUIRE(file.load(network, infilename) == Vector::DBC::Status::Ok);
 
-    /*
-     * The DBC file contains two messages that are separated by a line with one space.
-     * In previous versions this wasn't interpreted as message separator.
-     * Hence the signals of the second message were appended to the first message.
-     * chomp was modified to also chop away spaces and tabs.
-     */
     Vector::DBC::Message & message = network.messages[1];
     BOOST_REQUIRE(message.id == 1);
     BOOST_CHECK(message.name == "Message_1");
     Vector::DBC::Signal & signal = message.signals["Signal_1"];
     BOOST_CHECK(signal.name == "Signal_1");
+
+    /*
+     * The second message is separated by a line with one space.
+     * In previous versions this wasn't interpreted as message separator.
+     * Hence the signals of the second message were appended to the first message.
+     * chomp was modified to also chop away spaces and tabs.
+     */
     message = network.messages[2];
     BOOST_REQUIRE(message.id == 2);
     BOOST_CHECK(message.name == "Message_2");
     signal = message.signals["Signal_2"];
     BOOST_CHECK(signal.name == "Signal_2");
+
+    /*
+     * The third message is not separated by a line at all.
+     * In previous versions this wasn't interpreted as message separator.
+     * Hence the following message was assumed to be a signal too.
+     * The file::load method was modified to cope with this.
+     */
+    message = network.messages[3];
+    BOOST_REQUIRE(message.id == 3);
+    BOOST_CHECK(message.name == "Message_3");
+    signal = message.signals["Signal_3"];
+    BOOST_CHECK(signal.name == "Signal_3");
 
     /*
      * In Value Descriptions a description with leading space let to undefined behavior.