+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### [Fixed]
+- Allow to read files that have messages without blank line as separator
+
 ## [1.0.1] - 2016-05-20
 ### Added
 - Compiler hardening flags enabled